Crate hangman

Structs

GameData ğŸ”’

This is the Struct for the game The discovered letters lives and status are all printed

Enums

UserInputStatus ğŸ”’

This enum is used to tell a user if they have previously found letters It also keeps has the guessed and letters missed to work with the display

Functions

check_user_guess ğŸ”’

clear ğŸ”’

format_masked_string ğŸ”’

get_random_line ğŸ”’

main ğŸ”’

print_hangman ğŸ”’

I made changes to the hang man images below and added more lives because some of the phrases are very long

read_guess ğŸ”’

update_screen ğŸ”’

validate_user_guess ğŸ”’